+++ devices/control.js
@@ -11,23 +11,42 @@ control.prototype={
 
   createControlService(device, type){
     this.log.debug('adding new control')
-		let currentAmps=((device.maxChargingCurrent-32)*5/9).toFixed(2)
+		let currentAmps
+		if(this.platform.useFahrenheit){
+		currentAmps=((device.maxChargingCurrent-32)*5/9).toFixed(2)
+		}
+		else{
+			currentAmps=device.maxChargingCurrent
+		}
 		let controlService=new Service.Thermostat(type, device.id)
     controlService 
       .setCharacteristic(Characteristic.Name, type)
-      .setCharacteristic(Characteristic.StatusFault,false)
+      .setCharacteristic(Characteristic.StatusFault,Characteristic.StatusFault.NO_FAULT)
 			.setCharacteristic(Characteristic.TargetTemperature, currentAmps) //4.4444
 			.setCharacteristic(Characteristic.CurrentTemperature, currentAmps) 
-			.setCharacteristic(Characteristic.TemperatureDisplayUnits,false)
+			.setCharacteristic(Characteristic.TemperatureDisplayUnits,this.platform.useFahrenheit)
 			.setCharacteristic(Characteristic.TargetHeatingCoolingState,0) //off
 			.setCharacteristic(Characteristic.CurrentHeatingCoolingState,0)
     return controlService
   },
 
   configureControlService(device, controlService){
-		let min=-14.5
-		let max=4.5
-		if(device.maxAvailableCurrent==48){max=9}
+		let min
+		let max
+		let step
+		if(this.platform.useFahrenheit){
+			min=-14.5
+			max=4.5 
+			step=.5
+			if(device.maxAvailableCurrent==48){max=9}
+		}
+		else{
+			min=6
+			max=40
+			step=1
+			if(device.maxAvailableCurrent==48){max=48}
+		}
+		
     this.log.info("Configured %s control for %s" , controlService.getCharacteristic(Characteristic.Name).value, device.name)
 		controlService
       .getCharacteristic(Characteristic.TargetHeatingCoolingState)
@@ -40,16 +59,26 @@ control.prototype={
 		controlService
       .getCharacteristic(Characteristic.TargetTemperature)
 			.setProps({
-					minStep:.5,
+					minStep:step,
 					minValue:min,
 					maxValue:max
 			})
       .on('get', this.getControlAmps.bind(this, controlService))
       .on('set', this.setControlAmps.bind(this, device, controlService))
+		controlService
+      .getCharacteristic(Characteristic.TemperatureDisplayUnits)	
+			.on('get', this.getControlUnits.bind(this, controlService))
+      .on('set', this.setControlUnits.bind(this, device, controlService))
   },
 
 	setControlAmps(device, controlService, value, callback){
-		let amps=(value*1.8+32).toFixed(2)
+		let amps
+		if(this.platform.useFahrenheit){
+			amps=(value*1.8+32).toFixed(2)
+		}
+		else{
+			amps=value
+		}
 		this.log.debug('set amps',value, amps)
 		this.wallboxapi.setAmps(this.platform.token,device.id,amps).then(response=>{
 			this.log.debug(response.data)
@@ -60,13 +89,13 @@ control.prototype={
 	setControlState(device, controlService, value, callback){
 		this.wallboxapi.getChargerData(this.platform.token,device.id).then(response=>{
 			try{
-				locked=response.data.data.chargerData.locked
-				this.log.debug('check lock state = %s',locked)
+				connected=response.data.data.chargerData.status
+				this.log.debug('check connected state = %s',connected)
 			}catch(error){
-				locked=true
-				this.log.error("failed lock state check")
-			}			
-			if(!locked){
+				connected=209
+				this.log.error("failed connected state check")
+			}		
+			if(connected!=(161 && 209)){
 				this.log.debug('toggle switch state %s',controlService.getCharacteristic(Characteristic.Name).value)
 				if(controlService.getCharacteristic(Characteristic.StatusFault).value==Characteristic.StatusFault.GENERAL_FAULT){
 					callback('error')
@@ -104,13 +133,18 @@ control.prototype={
 				} 
 			}
 			else{
-				this.log.info('Charger must be unlocked for this operation')
-				controlService.getCharacteristic(Characteristic.TargetHeatingCoolingState).updateValue(!value)
+				this.log.info('Car must be connected for this operation')
+				controlService.getCharacteristic(Characteristic.On).updateValue(!value)
 				callback()
 			}	
 		})
-		//this.platform.startLiveUpdate(device)
 	},
+
+	setControlUnits(device, controlService, value, callback){
+		//this.platform.useFahrenheit=value
+		this.log.debug("change unit value")
+		callback()
+		},
 
 	getControlState(controlService, callback){
 		if(controlService.getCharacteristic(Characteristic.StatusFault).value==Characteristic.StatusFault.GENERAL_FAULT){
@@ -128,10 +162,26 @@ control.prototype={
 		}
 		else{
 			let currentValue=controlService.getCharacteristic(Characteristic.CurrentTemperature).value
-			if(currentValue>4.5){currentValue=4.44444}
+			if(this.platform.useFahrenheit){
+				if(currentValue>4.5){currentValue=4.44444}
+			}
+			else{
+				if(currentValue>40){currentValue=40}
+			}
 			callback(null, currentValue)
 		}
-	} 
+	}, 
+
+	getControlUnits(controlService, callback){
+		if(controlService.getCharacteristic(Characteristic.StatusFault).value==Characteristic.StatusFault.GENERAL_FAULT){
+			callback('error')
+		}
+		else{
+			let currentValue=controlService.getCharacteristic(Characteristic.TemperatureDisplayUnits).value
+			this.platform.useFahrenheit=currentValue
+			callback(null, currentValue)
+		}
+	}
 
 }