+++ devices/outlet.js
@@ -32,13 +32,13 @@ basicOutlet.prototype={
   setOutletValue(device, outletService, value, callback){
 		this.wallboxapi.getChargerData(this.platform.token,device.id).then(response=>{
 			try{
-				locked=response.data.data.chargerData.locked
-				this.log.debug('check lock state = %s',locked)
+				connected=response.data.data.chargerData.status
+				this.log.debug('check connected state = %s',connected)
 			}catch(error){
-				locked=true
-				this.log.error("failed lock state check")
-			}			
-			if(!locked){
+				connected=209
+				this.log.error("failed connected state check")
+			}		
+			if(connected!=(161 && 209)){
 				this.log.debug('toggle outlet state %s',outletService.getCharacteristic(Characteristic.Name).value)
 				if(outletService.getCharacteristic(Characteristic.StatusFault).value==Characteristic.StatusFault.GENERAL_FAULT){
 					callback('error')
@@ -78,12 +78,11 @@ basicOutlet.prototype={
 				} 
 			}
 			else{
-				this.log.info('Charger must be unlocked for this operation')
+				this.log.info('Car must be connected for this operation')
 				outletService.getCharacteristic(Characteristic.On).updateValue(!value)
 				callback()
 			}	
 		})
-		//this.platform.startLiveUpdate(device)
   },
 
 	getOutletValue(outletService, callback){